refactor: quotes hook (#1718)
* add use quote result

* move best quote

* add reset quote

* update earn provider

* update leverage provider

* add use tx review

* rename to use prepare tx review

* add biome json to avoid vscode formatting issues

* use format percentage

* remove obsolete spread syntax
