+++ biome.json
@@ -1,0 +1,8 @@
+{
+  "linter": {
+    "enabled": false
+  },
+  "formatter": {
+    "enabled": false
+  }
+}
