+++ src/app/earn/provider.tsx
@@ -1,29 +1,24 @@
 'use client'
 
-import { getTokenByChainAndSymbol, isAddressEqual } from '@indexcoop/tokenlists'
 import { useQuery } from '@tanstack/react-query'
 import {
   createContext,
   useCallback,
   useContext,
-  useEffect,
   useMemo,
   useState,
 } from 'react'
 import { isAddress } from 'viem'
 import { base } from 'viem/chains'
-import { usePublicClient } from 'wagmi'
 
 import { useQueryParams } from '@/app/earn/use-query-params'
 import { TransactionReview } from '@/components/swap/components/transaction-review/types'
 import { ETH, ICUSD, Token } from '@/constants/tokens'
 import { TokenBalance, useBalances } from '@/lib/hooks/use-balance'
-import { Quote, QuoteResult, QuoteType } from '@/lib/hooks/use-best-quote/types'
-import { getBestQuote } from '@/lib/hooks/use-best-quote/utils/best-quote'
-import { getFlashMintQuote } from '@/lib/hooks/use-best-quote/utils/flashmint'
-import { getIndexQuote } from '@/lib/hooks/use-best-quote/utils/index-quote'
+import { QuoteResult } from '@/lib/hooks/use-best-quote/types'
 import { useNetwork } from '@/lib/hooks/use-network'
-import { getTokenPrice, useNativeTokenPrice } from '@/lib/hooks/use-token-price'
+import { usePrepareTransactionReview } from '@/lib/hooks/use-prepare-transaction-review'
+import { useQuoteResult } from '@/lib/hooks/use-quote-result'
 import { useWallet } from '@/lib/hooks/use-wallet'
 import { isValidTokenInput, parseUnits } from '@/lib/utils'
 import {
@@ -98,9 +93,7 @@ const defaultParams = {
 }
 
 export function EarnProvider(props: { children: any }) {
-  const publicClient = usePublicClient()
   const { chainId: chainIdRaw } = useNetwork()
-  const nativeTokenPrice = useNativeTokenPrice(chainIdRaw)
   const { address } = useWallet()
   const {
     queryParams: { queryInputToken, queryOutputToken, queryIsMinting },
@@ -108,12 +101,6 @@ export function EarnProvider(props: { children: any }) {
   } = useQueryParams({ ...defaultParams, network: chainIdRaw })
 
   const [inputValue, setInputValue] = useState('')
-  const [quoteResult, setQuoteResult] = useState<QuoteResult>({
-    type: QuoteType.flashmint,
-    isAvailable: true,
-    quote: null,
-    error: null,
-  })
 
   const isMinting = queryIsMinting
   const inputToken = queryInputToken
@@ -141,6 +128,38 @@ export function EarnProvider(props: { children: any }) {
   const { balances, forceRefetchBalances } = useBalances(
     address,
     indexTokenAddresses,
+  )
+
+  const inputTokens = useMemo(() => {
+    if (isMinting) return getCurrencyTokens(chainId)
+    return indexTokens
+  }, [chainId, indexTokens, isMinting])
+
+  const outputTokens = useMemo(() => {
+    if (!isMinting) return getCurrencyTokens(chainId)
+    return indexTokens
+  }, [chainId, indexTokens, isMinting])
+
+  const inputTokenAmount = useMemo(
+    () =>
+      inputValue === ''
+        ? BigInt(0)
+        : parseUnits(inputValue, inputToken.decimals),
+    [inputToken, inputValue],
+  )
+
+  const { isFetchingQuote, quoteResult, resetQuote } = useQuoteResult({
+    address,
+    chainId,
+    isMinting,
+    inputToken,
+    outputToken,
+    inputTokenAmount,
+    inputValue,
+  })
+  const transactionReview = usePrepareTransactionReview(
+    isFetchingQuote,
+    quoteResult,
   )
 
   const {
@@ -208,33 +227,6 @@ export function EarnProvider(props: { children: any }) {
     },
   })
 
-  const inputTokenAmount = useMemo(
-    () =>
-      inputValue === ''
-        ? BigInt(0)
-        : parseUnits(inputValue, inputToken.decimals),
-    [inputToken, inputValue],
-  )
-
-  const inputTokens = useMemo(() => {
-    if (isMinting) return getCurrencyTokens(chainId)
-    return indexTokens
-  }, [chainId, indexTokens, isMinting])
-
-  const outputTokens = useMemo(() => {
-    if (!isMinting) return getCurrencyTokens(chainId)
-    return indexTokens
-  }, [chainId, indexTokens, isMinting])
-
-  const toggleIsMinting = useCallback(() => {
-    updateQueryParams({
-      isMinting,
-      inputToken: outputToken,
-      outputToken: inputToken,
-      network: chainId,
-    })
-  }, [chainId, isMinting, inputToken, outputToken, updateQueryParams])
-
   const onChangeInputTokenAmount = useCallback(
     (input: string) => {
       if (input === '') {
@@ -292,155 +284,18 @@ export function EarnProvider(props: { children: any }) {
 
   const reset = () => {
     setInputValue('')
-    setQuoteResult({
-      type: QuoteType.flashmint,
-      isAvailable: true,
-      quote: null,
-      error: null,
-    })
+    resetQuote()
     forceRefetchBalances()
   }
 
-  const fetchFlashMintQuote = async () => {
-    if (!address) return null
-    if (!chainId) return null
-    if (!publicClient) return null
-    if (inputTokenAmount <= 0) return null
-    if (!indexToken) return null
-    const inputTokenPrice = await getTokenPrice(inputToken, chainId)
-    const outputTokenPrice = await getTokenPrice(outputToken, chainId)
-    const isIcEth = isAddressEqual(
-      indexToken.address,
-      getTokenByChainAndSymbol(chainId, 'icETH')?.address,
-    )
-    return await getFlashMintQuote({
+  const toggleIsMinting = useCallback(() => {
+    updateQueryParams({
       isMinting,
-      account: address,
-      chainId,
-      inputToken,
-      inputTokenAmount: inputValue,
-      inputTokenAmountWei: inputTokenAmount,
-      inputTokenPrice,
-      outputToken,
-      outputTokenPrice,
-      slippage: isIcEth ? 0.5 : 0.1,
+      inputToken: outputToken,
+      outputToken: inputToken,
+      network: chainId,
     })
-  }
-
-  const { data: flashmintQuote, isFetching: isFetchingFlashMintQuote } =
-    useQuery({
-      queryKey: [
-        'flashmint-quote',
-        {
-          address,
-          chainId,
-          inputToken,
-          outputToken,
-          inputTokenAmount: inputTokenAmount.toString(),
-          publicClient,
-        },
-      ],
-      queryFn: fetchFlashMintQuote,
-      enabled:
-        !!address &&
-        !!chainId &&
-        !!inputToken &&
-        !!outputToken &&
-        !!publicClient &&
-        inputTokenAmount > 0, // Condition to trigger
-    })
-
-  const fetchSwapQuote = async () => {
-    if (!address) return null
-    if (!chainId) return null
-    if (!publicClient) return null
-    if (inputTokenAmount <= 0) return null
-    if (!indexToken) return null
-    const inputTokenPrice = await getTokenPrice(inputToken, chainId)
-    const outputTokenPrice = await getTokenPrice(outputToken, chainId)
-    return await getIndexQuote({
-      isMinting,
-      chainId,
-      address,
-      inputToken,
-      inputTokenAmount: inputValue,
-      inputTokenPrice,
-      outputToken,
-      outputTokenPrice,
-      nativeTokenPrice,
-      slippage: 0.1,
-    })
-  }
-
-  const { data: swapQuote, isFetching: isFetchingSwapQuote } = useQuery({
-    queryKey: [
-      'swap-quote',
-      {
-        address,
-        chainId,
-        inputToken,
-        outputToken,
-        inputTokenAmount: inputTokenAmount.toString(),
-        publicClient,
-      },
-    ],
-    queryFn: fetchSwapQuote,
-    enabled:
-      !!address &&
-      !!chainId &&
-      !!publicClient &&
-      !!inputToken &&
-      !!outputToken &&
-      inputTokenAmount > 0,
-  })
-
-  const isFetchingQuote = isFetchingFlashMintQuote || isFetchingSwapQuote
-
-  useEffect(() => {
-    const bestQuote = getBestYieldQuote(
-      flashmintQuote ?? null,
-      swapQuote ?? null,
-      chainId ?? -1,
-    )
-    setQuoteResult({
-      type: bestQuote?.type ?? QuoteType.flashmint,
-      isAvailable: true,
-      quote: bestQuote,
-      error: null,
-    })
-  }, [chainId, flashmintQuote, swapQuote])
-
-  const transactionReview = useMemo((): TransactionReview | null => {
-    if (isFetchingQuote || quoteResult === null) return null
-    const quote = quoteResult.quote
-    if (quote) {
-      return {
-        ...quote,
-        contractAddress: quote.contract,
-        quoteResults: {
-          bestQuote: QuoteType.flashmint,
-          results: {
-            flashmint: quoteResult,
-            index: null,
-            issuance: null,
-            redemption: null,
-          },
-        },
-        selectedQuote: QuoteType.flashmint,
-      }
-    }
-    return null
-  }, [isFetchingQuote, quoteResult])
-
-  useEffect(() => {
-    // Reset quotes
-    setQuoteResult({
-      type: QuoteType.flashmint,
-      isAvailable: true,
-      quote: null,
-      error: null,
-    })
-  }, [chainId, queryIsMinting, queryInputToken, queryOutputToken])
+  }, [chainId, isMinting, inputToken, outputToken, updateQueryParams])
 
   return (
     <EarnContext.Provider
@@ -475,37 +330,4 @@ export function EarnProvider(props: { children: any }) {
       {props.children}
     </EarnContext.Provider>
   )
-}
-
-function getBestYieldQuote(
-  flashmintQuote: Quote | null,
-  swapQuote: Quote | null,
-  chainId: number,
-): Quote | null {
-  if (!flashmintQuote && swapQuote) return swapQuote
-  if (flashmintQuote && !swapQuote) return flashmintQuote
-  if (
-    flashmintQuote &&
-    flashmintQuote.chainId !== chainId &&
-    swapQuote &&
-    swapQuote.chainId === chainId
-  )
-    return swapQuote
-  if (
-    flashmintQuote &&
-    flashmintQuote.chainId === chainId &&
-    swapQuote &&
-    swapQuote.chainId !== chainId
-  )
-    return flashmintQuote
-  if (flashmintQuote && swapQuote) {
-    const bestQuoteType = getBestQuote(
-      swapQuote.fullCostsInUsd,
-      flashmintQuote.fullCostsInUsd,
-      swapQuote.outputTokenAmountUsdAfterFees,
-      flashmintQuote.outputTokenAmountUsdAfterFees,
-    )
-    return bestQuoteType === QuoteType.index ? swapQuote : flashmintQuote
-  }
-  return null
 }
